Extract HeaderComponent
